+++ src/lib/InputInternal.tsx
@@ -1,9 +1,9 @@
-import { FieldValues, useFormContext, get, FieldError } from "react-hook-form";
+import { FieldValues, get, FieldError } from "react-hook-form";
 import { Input } from "reactstrap";
 import { useSafeNameId } from "src/lib/hooks/useSafeNameId";
 import { InputProps } from "./Input";
 import { useMarkOnFocusHandler } from "./hooks/useMarkOnFocusHandler";
-import { useInternalFormContext } from "./context/InternalFormContext";
+import { useFormContext } from "./context/FormContext";
 
 // This is two random guids concatenated. It is used to set the value of the option to undefined.
 const UNDEFINED_OPTION_VALUE = "CABB7A27DB754DA58C89D43ADB03FE0EC5EE3E25A6624D749F35CF2E92CFA784";
@@ -31,6 +31,7 @@ const InputInternal = <T extends FieldValues>(props: InputProps<T>) => {
   const {
     register,
     formState: { errors },
+    disabled: formDisabled,
   } = useFormContext();
 
   const { ref, ...rest } = register(name, {
@@ -39,7 +40,6 @@ const InputInternal = <T extends FieldValues>(props: InputProps<T>) => {
 
   const fieldError = get(errors, name) as FieldError | undefined;
   const hasError = !!fieldError;
-  const { disabled: formDisabled } = useInternalFormContext<T>();
 
   return (
     <>